changes towards revamp of mammotion integration
